Module Relude

module Array = Relude_Array;

`Relude` is the root namespace module where all of the other modules are exposed in a convenient hierarchy.

module ArrayZipper = Relude_ArrayZipper;

module AsyncData = Relude_AsyncData;

module AsyncResult = Relude_AsyncResult;

module Bool = Relude_Bool;

module Cont = Relude_ContT.Cont;

module ContT = Relude_ContT;

module Debounce = Relude_Debounce;

module Decimal = Relude_Decimal;

module Eq = Relude_Eq;

module Extensions = Relude_Extensions;

module Float = Relude_Float;

module Free = Relude_Free;

module Function = Relude_Function;

module Globals = Relude_Globals;

module HList = Relude_HList;

module HMap = Relude_HMap;

module Identity = Relude_Identity;

module Int = Relude_Int;

module Interface = Relude_Interface;

module IO = Relude_IO;

module Ior = Relude_Ior;

module Js = Relude_Js;

module List = Relude_List;

module ListZipper = Relude_ListZipper;

module Map = Relude_Map;

module Nea = Relude_NonEmpty.Array;

module Nel = Relude_NonEmpty.List;

module NonEmpty = Relude_NonEmpty;

module NonEmptyArray = Relude_NonEmpty.Array;

module NonEmptyList = Relude_NonEmpty.List;

module Option = Relude_Option;

module OptionT = Relude_OptionT;

module Ord = Relude_Ord;

module Ordering = Relude_Ordering;

module Reader = Relude_ReaderT.Reader;

module ReaderT = Relude_ReaderT;

module Result = Relude_Result;

module ResultT = Relude_ResultT;

module RIO = Relude_RIO;

module RWST = Relude_RWST;

module Sequence = Relude_Sequence;

module SequenceZipper = Relude_SequenceZipper;

module Set = Relude_Set;

module State = Relude_StateT.State;

module StateT = Relude_StateT;

module String = Relude_String;

module StringMap = Relude_StringMap;

module Throttle = Relude_Throttle;

module Timer = Relude_Timer;

module Tree = Relude_Tree;

module TreeZipper = Relude_TreeZipper;

module Tuple = Relude_Tuple;

module Tuple2 = Relude_Tuple2;

module Tuple3 = Relude_Tuple3;

module Tuple4 = Relude_Tuple4;

module Tuple5 = Relude_Tuple5;

module Unit = Relude_Unit;

module Unsafe = Relude_Unsafe;

module Validation = Relude_Validation;

module Void = Relude_Void;

module Writer = Relude_WriterT.Writer;

module WriterLog = Relude_WriterT.WriterLog;

module WriterT = Relude_WriterT;